Fix build and rename
